Add MemPass, move all shared functions to it.
